requirements.txt
flask
flask-socketio
redis
python-dotenv
eventlet
gunicorn
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
email-validator # Needed by Flask-WTF for email fields
Werkzeug # Explicitly listed, though often a Flask dependency